requirements.txt
asyncio
openai
emoji
discordpy-self==2.0.0
python-dotenv 